refactor: move main to cli module
