+++ src/aspeak/cli/__init__.py
@@ -1,1 +1,1 @@
-from .parser import parser
+from .main import main
